requirements.txt
flask[async]
gunicorn[gthread]
numpy
pandas
requests
aiohttp
multiprocess
scikit-learn
python-dotenv
joblib